Update README steps for make
Including the other couple of steps to make for others for whom it is not obvious. Generally people find out about GLEW when using it as dependency of another lib, and they may or may not be familiar with the make process. Leaving it vague leads to more people reporting issues, which I'm aiming to avoid here!