Fixed a bug with shell script and added gitattributes to prevent non-binary pulls
Cygwin was unable to filter the specs files using the blacklist due to
the argument in find was -name *.txt instead of -name "*.txt" .

Also, it is currently possible to pull under windows to result in CRLF
line endings in scripts and other source files that destroy the build.
.gitattributes were added to prevent non-binary pulls.

Both issues are fixed with this pull request.