Fixed a bug with shell script and added .gitattributes
Update_ext.sh does not utilize the blacklist correctly with cygwin due
to the argument -name in find is not quoted. Argument *.txt has been
changed to "*.txt" to fix the issue.

Also, it is currently possible for windows users to pull the repo in
CRLF format which destroys the scripts and source files. .gitattributes
were added to prevent this issue.