Support custom callback to resolve OpenGL entry points
There's a need to be able to define a custom callback to resolve
the OpenGL entry points. For example when integrating with 3rd party
toolkits such as Qt. Glew doesn't know the underlying implementation
and the entry point resolution really needs to happen through the
context object created by Qt.

Implementation for issue #106