Fix perl @INC for scripts which include make.pl
In perl 5.26, and some earlier perls that have been proactively
patched (Debian's perl 5.24, for example), '.' was removed from @INC
for security reasons.  For details, see:
  http://blogs.perl.org/users/todd_rinaldo/2016/11/how-removing-from-inc-is-about-to-break-cpan.html
  https://www.masteringperl.org/2017/01/perl-v5-26-removes-from-inc-but-dont-think-youre-safe/
et cetera.

This breaks the generation scripts which source make.pl with
  do 'bin/make.pl'

Although this might be indicative that it's time to refactor these
scripts, I opted for the simplest possible fix for now: restoring '.'
to @INC in each script that depends on make.pl.