Add VERSION and SOVERSION number to shared libraries
With this commit, when building or installing `glew`, VERSION and
SOVERSION symlinks to the `glew` shared library are created in the
library installation folder if the platform supports them and the linker
supports so-names.
VERSION and SOVERSION specify, respectively, the build version and API
version of the shared library.

As an example, under macOS, the following symlinks will be created:
libGLEW.2.0.0.dylib                      (the shared library)
libGLEW.2.0.dylib -> libGLEW.2.0.0.dylib (VERSION symlink)
libGLEW.dylib -> libGLEW.2.0.dylib       (SOVERSION symlink)

Fixes #141